SetupWizard: move sentences and implementation into a subcontext
