+++ eZ/Bundle/EzPublishLegacyBundle/Features/Context/FeatureContext.php
@@ -1,8 +1,6 @@
 <?php
 /**
- * File containing the SetupWizardContext class.
- *
- * This class contains specific setup wizard feature context for Behat.
+ * File containing the FeatureContext class for Legacy Bundle.
  *
  * @copyright Copyright (C) 1999-2014 eZ Systems AS. All rights reserved.
  * @license http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
@@ -11,22 +9,14 @@
 
 namespace eZ\Bundle\EzPublishLegacyBundle\Features\Context;
 
-use eZ\Bundle\EzPublishLegacyBundle\Features\Context\LegacyInternalSentences;
+use eZ\Bundle\EzPublishLegacyBundle\Features\Context\SubContexts\SetupWizard;
 use EzSystems\BehatBundle\Features\Context\Browser\BrowserContext;
-use PHPUnit_Framework_Assert as Assertion;
-use Behat\Behat\Context\Step;
-use Behat\Gherkin\Node\TableNode;
 
 /**
- * Setup Wizard context.
+ * FeatureContext context.
  */
-class FeatureContext extends BrowserContext implements LegacyInternalSentences
+class FeatureContext extends BrowserContext
 {
-    /**
-     * @var array This var should have the association between title in setup and package
-     */
-    protected $packages = array();
-
     public function __construct( array $parameters )
     {
         parent::__construct( $parameters );
@@ -35,91 +25,7 @@ class FeatureContext extends BrowserContext implements LegacyInternalSentences
             "setup wizard" => "/ezsetup",
         );
 
-        $this->packages += array(
-            'ez publish demo site' => 'ezdemo_site',
-            'ez publish demo site (without demo content)' => 'ezdemo_site_clean'
-        );
-    }
-
-    /**
-     * After this comment are the Legacy sentences implementation
-     *
-     * @see \eZ\Bundle\EzPublishLegacyBundle\Features\Context\LegacyInternalSentences
-     */
-
-    public function iAmOnStep( $stepTitle )
-    {
-        return array(
-            new Step\Then( 'I see "Setup Wizard" page' ),
-            new Step\Then( 'I see "' . $stepTitle . '" title' ),
-        );
-    }
-
-    public function iSelectPackage( $packageName, $version )
-    {
-        $package = $this->packages[strtolower( $packageName )];
-        Assertion::assertNotNull( $package, "Package '$packageName' not defined" );
-
-        // first select the package
-        $field = $this->getSession()->getPage()->findField( $package );
-        Assertion::assertNotNull( $field, "Couldn't find '$package' field" );
-        $this->browserFillField( $field );
-
-        // now verify version
-        // IMPORTANT: only verify the values shown on the page, does not actually
-        //      verify if the package is in a given version
-        $versionLabel = "$packageName (ver. $version)";
-        return array(
-            new Step\Then( 'I see "' . $versionLabel . '" title' )
-        );
-    }
-
-    public function iSeeImported( $packageName, $version )
-    {
-        $versionLabel = "$packageName (ver. $version)";
-        $packageXpath = "//h2[text() = '$versionLabel']";
-        $el = $this->getSession()->getPage()->find(
-            "xpath",
-            $packageXpath
-        );
-
-        Assertion::assertNotNull( $el, "Couldn't find '$versionLabel' package" );
-
-        $importElement = $this->findElmentAfterElement(
-            $this->findRow( $el )->findAll( "xpath", "td" ),
-            $packageXpath,
-            "//*[text() = 'Imported']"
-        );
-
-        Assertion::assertNotNull( $importElement, "Couldn't find 'Imported' for '$versionLabel' package" );
-    }
-
-    public function iSeeFollowingPackagesForVersionImported( $version, TableNode $packagesTable )
-    {
-        $packages = $this->convertTableToArrayOfData( $packagesTable );
-
-        foreach ( $packages as $packageName )
-        {
-            // this can't use the self::iSeeImported() since the versions don't
-            // have a space between "ver." and the actual version
-            $versionLabel = "$packageName (ver.{$version})";
-
-            // notice this xpath uses contains instead of the "=" because the
-            // text as an <enter> and trailing spaces, so it fails
-            $el = $this->getSession()->getPage()->find(
-                "xpath",
-                "//*[contains( text(), '$versionLabel' )]"
-            );
-
-            Assertion::assertNotNull( $el, "Couldn't find '$versionLabel' package" );
-
-            $importElement = $this->findElmentAfterElement(
-                $this->findRow( $el )->findAll( "xpath", "td" ),
-                "../*[contains( text(), '$versionLabel' )]",
-                "//*[text() = 'Imported']"
-            );
-
-            Assertion::assertNotNull( $importElement, "Couldn't find 'Imported' for '$versionLabel' package" );
-        }
+        // load sub contexts
+        $this->useContext( 'SetupWizard', new SetupWizard() );
     }
 }
